Extract contents of fragments as seed data
