+++ main-process/db/seeds/fragments.ts
@@ -1,0 +1,12 @@
+const content1 = `const hoge = 'hoge desu';
+const hoge2 = 'hoge desu desu';
+class Hoge {
+    hoge: any;
+    constructor(hoge) {
+        this.hoge = hoge;
+    }
+}`;
+const content2 = `puts 'Hello Ruby desu'
+puts 'Hello Ruby desu2'`;
+
+export { content1, content2 };
